refactor(container): unify structure and clean up styles
