+++ src/shared/ui/container/index.ts
@@ -1,1 +1,1 @@
-export { Container } from './ui'
+export { Container } from './container'
