--- src/widgets/header-main/ui.module.scss
@@ -1,115 +1,0 @@
-@import '../../shared/styles/variables.scss';
-@import '../../shared/styles/adaptive-px.scss';
-
-.header {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  height: $apx-66;
-  padding: $apx-14;
-  background: $background-light-gray;
-}
-
-.header__nav {
-  font-size: $apx-16;
-  width: $apx-1330;
-  height: $apx-38;
-  
-  display: flex;
-  align-items: center;
-  gap: $apx-35;
-}
-
-.header__logo {
-  width: $apx-116;
-  height: $apx-24;
-
-  svg{
-    width: 100%;
-    height: 100%;
-  }
-}
-
-.header__menu {
-  display: flex;
-  gap: $apx-30;
-
-  list-style-type: none;
-  margin-right: auto;
-}
-
-.header__link {
-  text-decoration: none;
-  color: $text-primary;
-  transition: 0.2s;
-
-  &:hover {
-    color: $text-heading;
-  }
-}
-
-.header__linkActive {
-  color: $text-heading;
-}
-
-.header__searchContainer {
-  position: relative;
-
-  svg {
-    position: absolute;
-    right: $apx-14;
-    top: 50%;
-    transform: translateY(-50%);
-
-    width: $apx-18;
-    height: $apx-18;
-  }
-
-	&:focus-within svg path {
-		stroke: $color-primary;
-	}
-}
-
-.header__search {
-	font-weight: 400;
-	font-size: $apx-16;
-
-  width: $apx-400;
-  height: $apx-38;
-  padding: 0 $apx-41 0 $apx-15;
-  border-radius: $apx-300;
-
-	border: $apx-2 solid rgb(42, 127, 255, 0);
-  outline: none;
-
-	background: $background-dirty-gray;
-	color: $text-heading;
-
-	&:focus {
-		border: 2px solid rgb(42, 127, 255, 0.8);
-	}
-
-	&::placeholder {
-		color: $text-secondary;
-	}
-
-	&::-webkit-search-cancel-button {
-  	-webkit-appearance: none;
-	}
-}
-
-.header__profile {
-  width: $apx-38; 
-  height: $apx-38;
-  border-radius: $apx-36;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: $color-primary;
-
-  svg{
-    width: $apx-18; 
-    height: $apx-18;
-  }
-}
